+++ eras/dijkstra/impl/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 0.3.0.0
 
+* Add `getDijkstraScriptsProvided`
+* Add `MissingRequiredGuards` constructor to `DijkstraUtxowPredFailure`
+* Add `DijkstraUtxoEnv` and use it as `Environemnt` in `STS` instance of `UTXOW`
 * Add `ApplyTick` instance for `DijkstraEra`
 * Add `WrongNetworkInDirectDeposit` constructor to `DijkstraUtxoPredFailure`
 * Add `SubWrongNetworkInDirectDeposit` constructor to `DijkstraSubUtxoPredFailure`